feat. 각종 로직 예외 처리 추가

+++ src/server/service/post.service.ts
@@ -1,12 +1,14 @@
 import {getUniqueString} from '@/lib/random';
+import {ForbiddenError, NotFoundError, UnauthorizedError} from '@/server/Error';
 import prisma from '@/server/modules/prisma';
 import s3 from '@/server/modules/s3';
 import {PostRepository} from '@/server/repositories/post.repository';
-import {Role} from '@prisma/client';
+import {AuthService} from '@/server/service/auth.service';
 
 export class PostService {
   constructor(
     private readonly postRepository: PostRepository = new PostRepository(),
+    private readonly authService = new AuthService(),
     private readonly prismaHelper = prisma,
     private readonly s3Helper = s3,
   ) {}
@@ -56,7 +58,6 @@ export class PostService {
    */
   async getNoticePostList({categoryId}: {categoryId: number}) {
     const list = await this.postRepository.getNoticeList(categoryId);
-
     return {list};
   }
 
@@ -65,6 +66,8 @@ export class PostService {
    */
   async getPostDetail(id: number) {
     const post = await this.postRepository.getDetail(id);
+    if (!post) throw NotFoundError();
+
     return post;
   }
 
@@ -72,37 +75,24 @@ export class PostService {
    * 게시글 생성
    */
   async create(
-    user: {
-      id: number;
-      role: Role;
-    },
-    post: {
-      title: string;
-      content: string;
-
-      images: string[];
-      categoryId: number;
-      subCategoryId?: number;
-      isNotice: boolean;
-    },
+    title: string,
+    content: string,
+    images: string[],
+    categoryId: number,
+    subCategoryId?: number,
+    isNotice?: boolean,
   ) {
-    const {images, title, categoryId, subCategoryId, content, isNotice} = post;
-
-    if (
-      isNotice &&
-      user.role !== 'SUPER_ADMIN' &&
-      user.role !== 'BOARD_ADMIN'
-    ) {
-      throw new Error('404 Not Found');
-    }
+    const user = isNotice
+      ? await this.authService.verifyAdmin()
+      : await this.authService.verifyUser();
 
     const newPost = await this.postRepository.create(
       user.id,
       title,
       content,
-      isNotice,
       categoryId,
       subCategoryId,
+      isNotice,
     );
     // 이미지 파일 이동
     const movedImageUrls = await Promise.all(
@@ -133,6 +123,8 @@ export class PostService {
    * 게시글 좋아요
    */
   async likePost(postId: number, userId: number) {
+    await this.authService.verifyUserId(userId);
+
     const like = await this.prismaHelper.likeToPost.findUnique({
       where: {
         userId_postId: {
@@ -166,6 +158,8 @@ export class PostService {
    * 게시글 싫어요
    */
   async dislikePost(postId: number, userId: number) {
+    await this.authService.verifyUserId(userId);
+
     const dislikes = await this.prismaHelper.dislikeToPost.findUnique({
       where: {userId_postId: {userId, postId}},
     });
@@ -206,6 +200,16 @@ export class PostService {
    * 게시글 삭제
    */
   async delete(postId: number, userId: number) {
+    const loginUser = await this.authService.verifyUser();
+
+    if (!loginUser) {
+      throw UnauthorizedError();
+    }
+
+    if (loginUser.id !== userId && loginUser.role !== 'SUPER_ADMIN') {
+      throw ForbiddenError();
+    }
+
     await this.postRepository.delete(postId, userId);
   }
